requirements.txt
Flask>=3.0
anthropic>=0.40.0
pdfplumber>=0.11
python-dotenv>=1.0
Werkzeug>=3.0
# Optional — improves skill extraction. Falls back to regex if not installed.
# spacy>=3.7
